STAR-515 Add guardrail for SASI and STAR-512 Add guardrails for total and per-table indexes for SAI (#140)
* STAR-512 Add guardrails for total and per-table indexes for SAI


* STAR-515 Add guardrail for SASI


(cherry picked from commit e2ae18ead10d7186a76bb1a8299bfcd4ddf3820e)
(cherry picked from commit eaf8c59b3820a575742370aa6b4936d3783bc49b)
(cherry picked from commit 753fb6916ba32673a479b8c8e39cec447aa97012)
(cherry picked from commit 0fafd7b053fc2520219ab8a0243362539696baf8)
(cherry picked from commit 41d26d3e2dc86f8be5e1639ac68d684eb3e57772)
(cherry picked from commit 3a40eae385b31cac03b61cb81fd5042c0f93964c)

+++ test/unit/org/apache/cassandra/guardrails/GuardrailsOnTableTest.java
@@ -44,7 +44,7 @@ public class GuardrailsOnTableTest extends GuardrailTester
     private static final String ALTER_VIEW = "ALTER MATERIALIZED VIEW %s.%s WITH %s";
     private static long defaultTablesSoftLimit;
     private static long defaultTableHardLimit;
-    private static long defaultMVPerTableFailureThreshold;
+    private static int defaultMVPerTableFailureThreshold;
     private static Set<String> defaultTablePropertiesDisallowed;
     private static Set<String> defaultTablePropertiesIgnored;
